pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "workflow-eval-env"
version = "1.0.0"
description = "Production-grade AI evaluation environment for professional workflows"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
baseline = ["openai>=1.0.0"]
dev = ["pytest>=7.0", "ruff>=0.1.0"]

[tool.setuptools.packages.find]
include = ["tasks*", "grading*", "baseline*"]
